redirect home (/) to /rpcs
