+++ admin/web/next.config.js
@@ -1,0 +1,14 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  async redirects() {
+    return [
+      {
+        source: '/',
+        destination: '/rpcs',
+        permanent: false,
+      },
+    ];
+  },
+};
+
+module.exports = nextConfig;
